+++ src/shared/styles/global.css
@@ -58,10 +58,12 @@
 
 /* Brand and UI color palette */
 :root {
+  --main: #90212a;
   --corkcharge-gray: #9fa2aa;
   --gray-1: #f3f3f6;
   --gray-4: #c5c8cf;
   --gray-6: #80818b;
+  --gray-8: #35353f;
 }
 
 @layer components {
